fix, logging in is now a choice and not a must
