+++ package.json
@@ -41,7 +41,6 @@
     "react-native-reanimated": "^2.14.1",
     "react-native-safe-area-context": "^4.4.1",
     "react-native-screens": "^3.18.2",
-    "react-native-vector-icons": "^9.2.0",
     "yarn": "^1.22.19"
   },
   "devDependencies": {
